moved test unyt_array compare fixutre to utils
