+++ tests/params/test_freestream.py
@@ -33,6 +33,11 @@ def test_freesteam():
 
     to_file_from_file_test(fs)
 
+    p = fl.Flow360Params(freestream={"Mach": 1, "temperature": 300, "mu_ref": 1})
+    p = fl.Flow360Params(freestream={"Mach": 0, "Mach_ref": 1, "temperature": 300, "mu_ref": 1})
+    with fl.SI_unit_system:
+        p = fl.Flow360Params(freestream={"velocity": 1})
+
     with pytest.raises(pd.ValidationError):
         fs = FreestreamFromMach(Mach=-1, Temperature=100)
